+++ Vehicles.API/Data/Entities/History.cs
@@ -43,14 +43,14 @@ namespace Vehicles.API.Data.Entities
 
         [Display(Name = "Total Mano de Obra")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
-        public decimal TotalLabor => Details == null ? 0 : Details.Sum(x => x.LaborPrice);
+        public int TotalLabor => Details == null ? 0 : Details.Sum(x => x.LaborPrice);
 
         [Display(Name = "Total Repuestos")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
-        public decimal TotalSpareParts => Details == null ? 0 : Details.Sum(x => x.SparePartsPrice);
+        public int TotalSpareParts => Details == null ? 0 : Details.Sum(x => x.SparePartsPrice);
 
         [Display(Name = "Total")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
-        public decimal Total => Details == null ? 0 : Details.Sum(x => x.TotalPrice);
+        public int Total => Details == null ? 0 : Details.Sum(x => x.TotalPrice);
     }
 }
