+++ Vehicles.API/Models/DetailViewModel.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using Vehicles.API.Data.Entities;
 
 namespace Vehicles.API.Models
 {
@@ -13,12 +12,12 @@ namespace Vehicles.API.Models
         [Display(Name = "Precio Mano de Obra")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
-        public decimal LaborPrice { get; set; }
+        public int LaborPrice { get; set; }
 
         [Display(Name = "Precio Repuestos")]
         [DisplayFormat(DataFormatString = "{0:C2}")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
-        public decimal SparePartsPrice { get; set; }
+        public int SparePartsPrice { get; set; }
 
         [Display(Name = "Observación")]
         [DataType(DataType.MultilineText)]
